Cancel ignored when linger is enabled

+++ libsubmit/libsubmit/providers/aws/aws.py
@@ -625,6 +625,10 @@ class AWSProvider(ExecutionProvider, RepresentationMixin):
             Each entry in the list will contain False if the operation fails. Otherwise, the entry will be True.
         """
 
+        if self.linger is True :
+            logger.debug("Ignoring cancel requests due to linger mode")
+            return [False for x in job_ids]
+
         try:
             self.client.terminate_instances(InstanceIds=list(job_ids))
         except Exception as e:
